Merge the scons-branch to trunk.
This basicly adds scons as a buildsystem. This does not mean that auto* is dropped instantly (the added files should not collide) but allows for some testing in the transition-process. The auto*-files are deleted later on...

Due to this, the config.h.in (which gets built by autogen.sh) is copied from config.h.in.scons using scons. When dropping make this file should be moved to config.h.in.

The scons-system is probably missing a lot of things you are used to in make. But at least here it does build a usable libffado. :-)

To use scons, simply do a "scons". If you want to know about configuration, "scons -h" allows you to view and set parameters, which are saved from there on. What do I mean with this? "scons -h PREFIX=$HOME" sets the install-prefix to $HOME and shows the helptext that already incorporates the changed option. Another "scons -h" will show the same help-text and scons remembers the changed prefix. "scons" now builds the libs and apps and "scons install" installs everything into the prefix. "scons -c install" cleans the local dir and uninstalls the package while "scons -c" just cleans the local dir.

BTW: I also set several more ignores, I am just to lazy to split these two commits. ;-)



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado/src/libstreaming@561 2be59082-3212-0410-8809-b0798e1608f0