Implement a BackLog. This is a buffer that stores the last debug messages, regardless of whether their debuglevel is high enough or not. It can be used to run programs at a low verbosity level, but print highly verbose output when something goes wrong.
The buffer is a ringbuffer where the oldest message is overwritten when a new
one is added. The readout however is FIFO, oldest messages first.

Use the debugShowBackLog() macro when you want to print the backlog to stderr.

Note that the backlog is not thread safe so sometimes there might be some errors
in the lines, but it otherwise would require the introduction of more locks in
the critical path.



git-svn-id: http://subversion.ffado.org/ffado/trunk@670 2be59082-3212-0410-8809-b0798e1608f0