Make the PanelManager a QWidget again.
Instead we introduce a new FFADOWindow that is the mainwindow. And we have this window appear before we try to connect to dbus which should make the look and feel a bit more consistent for non-console users. And the retry-question is not asked via QMessageBox anymore but with an in-gui widget. Which also obsoletes some predefined loops for retry...


git-svn-id: http://subversion.ffado.org/ffado/branches/libffado-2.0@1434 2be59082-3212-0410-8809-b0798e1608f0