Heavy refactoring.
I moved writeFunc and readFunc to .h and merged the two for-loops in
prepare(). This code is actually now in prepareSP(), prepare() only
contains the calling loops.

Note that in contrast to the previous code, I conditionally set dll_bw
depending whether it's RX or TX. I guess it was a copy-and-paste error
before. (I'm talking about line 1001 in r1936)

Works for my Alesis device.



git-svn-id: http://subversion.ffado.org/ffado/trunk@1937 2be59082-3212-0410-8809-b0798e1608f0