DICE: Introduce DICE_FL_OFFSET
This is supposed to be used in conjunction with readReg, writeReg and the
corresponding block functions.

They all use DICE_REGISTER_BASE as a constant offset, to which we have
to add an offset when talking to the firmware interface.

If you feel a minus is more readable than xor, feel free to change the
definition.

git-svn-id: http://subversion.ffado.org/ffado/trunk@2069 2be59082-3212-0410-8809-b0798e1608f0