DICE-EAP Use vector of pair rather than map for router config
By Philippe Carriere:

router config (m_routes2) is the image of what is really written in the flash.
The meaning of the router and peak space flash content is entirely
manufacturer dependent.

Then even the ordering followed by the destinations may be of importance
so it must be a priori preserved.

git-svn-id: http://subversion.ffado.org/ffado/trunk@2095 2be59082-3212-0410-8809-b0798e1608f0