[BeBoB/Yamaha] Add vendor specific command for BeBoB/Yamaha devices
Yamaha devices based on BeBoB chipset have vendor-specific AV/C commands. With these commands, the devices tell its driver whether to detect digital input or not, its clock is synchronized or not.

There is the strange because 'company id' field is different between commands. I guess this is due to some OEM relationships.

git-svn-id: http://subversion.ffado.org/ffado/trunk@2438 2be59082-3212-0410-8809-b0798e1608f0