[BeBoB/MAudio] update configulation file for BeBoB/MAudio devices
The entries for Ozonic/Firewire Solo/Firewire Audiophile are updated to use mixer GUI. The entry for Firewire 410 is newly added.

There is a note for Firewire Audiophile and Firewire 410. Just after powering on, these devices wait to download firmware. FFADO developers didn't implement this. If you use with firmware version 5058, the devices have firmware
image in its ROM and the driver can call it to be loaded by sending a cue.

The cue is '0x01000000 00001101 00000000' and should be sent to 0xffffc8021000 just ONE TIME.


git-svn-id: http://subversion.ffado.org/ffado/trunk@2448 2be59082-3212-0410-8809-b0798e1608f0