Support Saffire PRO 26.  Patch and commit message by Hector Martin.
Not to be confused with the older Pro 26 I/O, which ffado already
supports; the PRO 26 was released in 2014. This is basically a cross
between the 14, the 24, and the 40 of the same range. Two transmitters,
like the 40, registers offsets that are closer to the 14, and actual
capabilities are just like a 24 with two extra channels and some
differences in the monitoring config.

All streams and inputs have been tested, other than ADAT which I don't
have a test source for. ADAT-as-SPDIF has been tested, however.

I removed the mono switches in the monitoring panel, since those don't
seem to do anything on this model. This model also doesn't have any soft
switches for input settings, as all inputs are either fixed,
auto-switching, or have dedicated front panel buttons.

The only issue that I found is certain clocking instability with the
S/PDIF input when used in 96kHz mode, and it seems it also interacts
with ADAT-as-SPDIF in strange ways. This is likely a bug/limitation of
the interface and not a problem in libffado, though. I was testing
S/PDIF by looping it back, so it's also possible that the issue would
not be present with external S/PDIF sources.

Signed-off-by: Hector Martin <marcan@marcan.st>
Signed-off-by: Jonathan Woithe <jwoithe@just42.net>



git-svn-id: http://subversion.ffado.org/ffado/trunk@2590 2be59082-3212-0410-8809-b0798e1608f0